Block library: Extract `deprecated` field to their own files (p.2)
