+++ packages/block-library/src/paragraph/index.js
@@ -1,149 +1,32 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-import { isFinite, omit } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import {
-	RawHTML,
-} from '@wordpress/element';
-import {
-	getColorClassName,
-	RichText,
-} from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
+import deprecated from './deprecated';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
 
-const { name, attributes: schema } = metadata;
+const { name } = metadata;
 
 export { metadata, name };
-
-const supports = {
-	className: false,
-};
 
 export const settings = {
 	title: __( 'Paragraph' ),
 	description: __( 'Start with the building block of all narrative.' ),
 	icon,
 	keywords: [ __( 'text' ) ],
-	supports,
+	supports: {
+		className: false,
+	},
 	transforms,
-	deprecated: [
-		{
-			supports,
-			attributes: {
-				...schema,
-				width: {
-					type: 'string',
-				},
-			},
-			save( { attributes } ) {
-				const {
-					width,
-					align,
-					content,
-					dropCap,
-					backgroundColor,
-					textColor,
-					customBackgroundColor,
-					customTextColor,
-					fontSize,
-					customFontSize,
-				} = attributes;
-
-				const textClass = getColorClassName( 'color', textColor );
-				const backgroundClass = getColorClassName( 'background-color', backgroundColor );
-				const fontSizeClass = fontSize && `is-${ fontSize }-text`;
-
-				const className = classnames( {
-					[ `align${ width }` ]: width,
-					'has-background': backgroundColor || customBackgroundColor,
-					'has-drop-cap': dropCap,
-					[ fontSizeClass ]: fontSizeClass,
-					[ textClass ]: textClass,
-					[ backgroundClass ]: backgroundClass,
-				} );
-
-				const styles = {
-					backgroundColor: backgroundClass ? undefined : customBackgroundColor,
-					color: textClass ? undefined : customTextColor,
-					fontSize: fontSizeClass ? undefined : customFontSize,
-					textAlign: align,
-				};
-
-				return (
-					<RichText.Content
-						tagName="p"
-						style={ styles }
-						className={ className ? className : undefined }
-						value={ content }
-					/>
-				);
-			},
-		},
-		{
-			supports,
-			attributes: omit( {
-				...schema,
-				fontSize: {
-					type: 'number',
-				},
-			}, 'customFontSize', 'customTextColor', 'customBackgroundColor' ),
-			save( { attributes } ) {
-				const { width, align, content, dropCap, backgroundColor, textColor, fontSize } = attributes;
-				const className = classnames( {
-					[ `align${ width }` ]: width,
-					'has-background': backgroundColor,
-					'has-drop-cap': dropCap,
-				} );
-				const styles = {
-					backgroundColor,
-					color: textColor,
-					fontSize,
-					textAlign: align,
-				};
-
-				return <p style={ styles } className={ className ? className : undefined }>{ content }</p>;
-			},
-			migrate( attributes ) {
-				return omit( {
-					...attributes,
-					customFontSize: isFinite( attributes.fontSize ) ? attributes.fontSize : undefined,
-					customTextColor: attributes.textColor && '#' === attributes.textColor[ 0 ] ? attributes.textColor : undefined,
-					customBackgroundColor: attributes.backgroundColor && '#' === attributes.backgroundColor[ 0 ] ? attributes.backgroundColor : undefined,
-				}, [ 'fontSize', 'textColor', 'backgroundColor' ] );
-			},
-		},
-		{
-			supports,
-			attributes: {
-				...schema,
-				content: {
-					type: 'string',
-					source: 'html',
-					default: '',
-				},
-			},
-			save( { attributes } ) {
-				return <RawHTML>{ attributes.content }</RawHTML>;
-			},
-			migrate( attributes ) {
-				return attributes;
-			},
-		},
-	],
+	deprecated,
 	merge( attributes, attributesToMerge ) {
 		return {
 			content: ( attributes.content || '' ) + ( attributesToMerge.content || '' ),
