+++ packages/block-library/src/pullquote/deprecated.js
@@ -1,0 +1,74 @@
+/**
+ * WordPress dependencies
+ */
+import { RichText } from '@wordpress/block-editor';
+
+const blockAttributes = {
+	value: {
+		type: 'string',
+		source: 'html',
+		selector: 'blockquote',
+		multiline: 'p',
+	},
+	citation: {
+		type: 'string',
+		source: 'html',
+		selector: 'cite',
+		default: '',
+	},
+	mainColor: {
+		type: 'string',
+	},
+	customMainColor: {
+		type: 'string',
+	},
+	textColor: {
+		type: 'string',
+	},
+	customTextColor: {
+		type: 'string',
+	},
+};
+
+const deprecated = [
+	{
+		attributes: {
+			...blockAttributes,
+		},
+		save( { attributes } ) {
+			const { value, citation } = attributes;
+			return (
+				<blockquote>
+					<RichText.Content value={ value } multiline />
+					{ ! RichText.isEmpty( citation ) && <RichText.Content tagName="cite" value={ citation } /> }
+				</blockquote>
+			);
+		},
+	}, {
+		attributes: {
+			...blockAttributes,
+			citation: {
+				type: 'string',
+				source: 'html',
+				selector: 'footer',
+			},
+			align: {
+				type: 'string',
+				default: 'none',
+			},
+		},
+
+		save( { attributes } ) {
+			const { value, citation, align } = attributes;
+
+			return (
+				<blockquote className={ `align${ align }` }>
+					<RichText.Content value={ value } multiline />
+					{ ! RichText.isEmpty( citation ) && <RichText.Content tagName="footer" value={ citation } /> }
+				</blockquote>
+			);
+		},
+	},
+];
+
+export default deprecated;
